Add another language example (merge sort).
For merge sort to work, lists need to be a thing. Right now, those are defined by bookending the list items with "start list of [type]" and "end list." A few list functions are also needed: count, empty, first, rest, take, drop, and prepend make an appearance; append, take nth, drop nth, and others would be good candidates to include in the language as well.

This is sort of a fancy merge sort in that it takes a comparator instead of just assuming less than. That's mostly to illustrate/test calling variables as functions. The added verbosity necessitated by passing the comparator around suggests that something like Clojure's `let` together with a partial application function might be useful.