Add a more complex language example (a Brainfuck interpreter).
With this example, the core language should be pretty much done. There are two new collection types, maps and sets, some new standard functions, and string/character/symbol types. Provided that the example is actually correct (it's kind of late and there's kind of a lot of code, so who knows), the language is Turing complete.

The names for the collection functions are taken from Immutable.js. The Immutable API is generally really nice, and running with those names seems more reasonable than imposing traditional Lispy names... `cons`, `rest`, `butlast`, et al. aren't super descriptive. Also, that library is probably going to be the basis for the implementation anyway.

As for the Brainfuck interpreter, it's not going to win any awards. The memory is backed by a map, meaning that the pointer can venture into negative numbers, and cell size is left up to the implementation. The quick-and-dirty Clojure version that served as a reference for this example blows stack on pretty much every non-trivial example, suggesting that the algorithm blows in the pejorative sense as well.

One notable outstanding issue: it's currently impossible to call an anonymous function. Given the current syntax, there doesn't seem to be an obvious easy way to fix that, but it's probably worth some more thought.