Clean up the Brainfuck example.
Amazingly, the example didn't look all that broken. None of the maps were closed; that was probably the dumbest mistake. There was also a mistake in the parse function (`unshift` instead of `push`) as well as a missed opportunity to use an actual stack.

The example is a little bit shorter thanks to the elimination of the `increment` and `decrement` functions. Those functions definitely belong in the standard library, and the return-zero-if-null stuff definitely belongs in the `read` function.

Input handling is still kind of sloppy, and that's maybe something to address later on.