Start work on the type system.
The main goal of the type system is to abstract over any implementation details—it shouldn't matter that the language is implemented in JavaScript. That goal is most evident in the Integer type, which uses the big-integer library to allow integers of any size. I'm not totally sure what to do about other numeric types yet, but I'm leaning toward a Float type and something like Clojure's Ratio type.

The implementation uses Immutable.js for collections, and all of the primitive types are Immutable.js value objects, so that should make the language implementation pretty straightforward.