Alter JSDoc annotations for math functions.
I'm still waffling on the best way to document stuff, so this is definitely subject to change again. It'd probably be easier to just use ESDoc, but there are already some cases it can't handle. For instance, the Character/Integer/String classes don't use class syntax (to match the Immutable collections), so ESDoc can't document them as classes (AFAIK).